Reorganizing the routes & fixing a bug with benchmarks
