+++ routes/groupRoutes.js
@@ -1,6 +1,5 @@
 const userHandler = require(`../handlers/userHandler`);
 const groupHandler = require(`../handlers/groupHandler`);
-const bookHandler = require(`../handlers/bookHandler`);
 const postHandler = require(`../handlers/postHandler`);
 
 
@@ -18,7 +17,7 @@ module.exports = app => {
     });
 
     //Need to find a way for this to approve the user to the group
-    app.put(`/api/adduser`, userHandler.isLoggedIn, async (req, res) => {
+    app.put(`/api/addusertogroup`, userHandler.isLoggedIn, async (req, res) => {
         const { addedUserID, groupID } = req.body;
         const isMod = await groupHandler.checkGroupMod(req.user._id, groupID);
         if (isMod) {
@@ -28,13 +27,6 @@ module.exports = app => {
             //TODO Need to have some sort of display on the front end 
             return "You need to be a moderator to add users to the group"
         };
-    });
-
-    app.get(`/api/searchbook/:book`, async (req, res) => {
-        const searchedBook = req.params.book;
-
-        const response = await bookHandler.searchGoogleBook(searchedBook);
-        res.json(response);
     });
 
     //Before post gets to here validate that there isn't a blank field
@@ -57,25 +49,6 @@ module.exports = app => {
         const groupPosts = await postHandler.getGroupPost(groupID);
         const sortedPosts = await postHandler.sortPostByDate(groupPosts);
         res.json(sortedPosts);
-    });
-
-    app.post(`/api/addbook`, userHandler.isLoggedIn, async (req, res) => {
-        //When the user picks a book from google books, this takes the data and saves it down
-        const { chosenBook, groupID } = req.body;
-
-        const bookForGroup = await bookHandler.queryAndSaveToDB(chosenBook);
-
-        //Checks if the user is a mod of the group they're currently trying to update
-        const isMod = await groupHandler.checkGroupMod(req.user._id, groupID);
-        if (isMod) {
-            //This function will move the current book into the past book and add the book in chosenBook to the current book
-            const updatedGroup = await bookHandler.updateCurrentBook(bookForGroup._id, groupID);
-            res.json(updatedGroup);
-        } else {
-            //TODO Add something to show if they tried to update a group they weren't a mod for
-            res.json({ 'error': `Moderator needed to update book` });
-            return;
-        }
     });
 
     //Adds the amount of pages or chapters to the Club
